requirements.txt
instabot==0.73.0
python-dotenv==0.10.3
